Introduce IOLoop, make DataReader/DataWriter implement IOHandler, introduce TransferHandler + small refactoring + TransferActor WIP
