+++ volga/streaming/runtime/network/buffer/test_buffer_serialization.py
@@ -2,8 +2,8 @@ import unittest
 
 import simplejson
 
-from volga.streaming.runtime.transfer.buffer import serialize, get_buffer_id, get_channel_id, get_payload
-from volga.streaming.runtime.transfer.utils import bytes_to_str, int_to_bytes, bytes_to_int, str_to_bytes
+from volga.streaming.runtime.network.buffer import serialize, get_buffer_id, get_channel_id, get_payload
+from volga.streaming.runtime.network.utils import bytes_to_str, int_to_bytes, bytes_to_int, str_to_bytes
 
 
 class TestBufferSerialization(unittest.TestCase):
