+++ volga/streaming/runtime/network/experimental/asynch/data_reader_async.py
@@ -1,3 +1,4 @@
+import asyncio
 import logging
 import time
 from collections import deque
@@ -5,24 +6,24 @@ from typing import List, Optional
 
 import zmq
 
-from volga.streaming.runtime.transfer.channel import Channel, ChannelMessage, LocalChannel, RemoteChannel
-
+import zmq.asyncio as zmq_async
 import simplejson
 
-from volga.streaming.runtime.transfer.buffer import AckMessage, get_buffer_id, get_payload, AckMessageBatch
-from volga.streaming.runtime.transfer.experimental.threaded import DataHandlerBase
-from volga.streaming.runtime.transfer.utils import bytes_to_str
+from volga.streaming.runtime.network.buffer.buffer import Buffer, AckMessage, get_buffer_id, get_payload
+from volga.streaming.runtime.network.experimental.asynch.async_data_handler_base import AsyncDataHandlerBase
+from volga.streaming.runtime.network.experimental.channel import BiChannel, LocalBiChannel, RemoteBiChannel
+from volga.streaming.runtime.network.utils import bytes_to_str
 
 logger = logging.getLogger("ray")
 
 
-class DataReaderV2(DataHandlerBase):
+class DataReaderV2Async(AsyncDataHandlerBase):
     def __init__(
         self,
         name: str,
-        channels: List[Channel],
+        channels: List[BiChannel],
         node_id: str,
-        zmq_ctx: zmq.Context,
+        zmq_ctx: zmq_async.Context,
     ):
         super().__init__(
             name=name,
@@ -32,10 +33,9 @@ class DataReaderV2(DataHandlerBase):
         )
 
         self._channels = channels
-        self._channel_map = { c for c in self._channels}
+        self._channel_map = {c.channel_id: c for c in self._channels}
 
         self._buffer_queue = deque()
-        self._acks_queues = {c.channel_id: deque() for c in self._channels}
 
     def _init_send_sockets(self):
         for channel in self._channels:
@@ -45,13 +45,9 @@ class DataReaderV2(DataHandlerBase):
             # TODO set HWM
             send_socket = self._zmq_ctx.socket(zmq.PUSH)
             send_socket.setsockopt(zmq.LINGER, 0)
-            send_socket.setsockopt(zmq.SNDHWM, 10000)
-            send_socket.setsockopt(zmq.RCVHWM, 10000)
-            send_socket.setsockopt(zmq.SNDBUF, 2000 * 1024)
-            send_socket.setsockopt(zmq.RCVBUF, 2000 * 1024)
-            if isinstance(channel, LocalChannel):
+            if isinstance(channel, LocalBiChannel):
                 send_socket.bind(channel.ipc_addr_in)
-            elif isinstance(channel, RemoteChannel):
+            elif isinstance(channel, RemoteBiChannel):
                 raise ValueError('RemoteChannel not supported yet')
             else:
                 raise ValueError('Unknown channel type')
@@ -65,10 +61,6 @@ class DataReaderV2(DataHandlerBase):
             # TODO set HWM
             rcv_socket = self._zmq_ctx.socket(zmq.PULL)
             rcv_socket.setsockopt(zmq.LINGER, 0)
-            rcv_socket.setsockopt(zmq.SNDHWM, 10000)
-            rcv_socket.setsockopt(zmq.RCVHWM, 10000)
-            rcv_socket.setsockopt(zmq.SNDBUF, 2000 * 1024)
-            rcv_socket.setsockopt(zmq.RCVBUF, 2000 * 1024)
             if isinstance(channel, LocalChannel):
                 rcv_socket.connect(channel.ipc_addr_out)
             elif isinstance(channel, RemoteChannel):
@@ -90,58 +82,47 @@ class DataReaderV2(DataHandlerBase):
         msg = simplejson.loads(bytes_to_str(data))
         return msg
 
-    # def _start_sender_loop(self):
-    #     pass
-
-    def _receive_loop(self):
+    async def _receive_loop(self):
         print('reader rcv loop started')
         while self.running:
-            sockets_and_flags = self._rcv_poller.poll()
+            sockets_and_flags = await self._rcv_poller.poll()
             for (socket, _) in sockets_and_flags:
                 # TODO check buffer pool capacity
                 # TODO limit number in-flight reads, indicate channel backpressure if read times-out
-                self._read(socket)
+                asyncio.create_task(self._read(socket))
 
-    def _read(self, rcv_socket: zmq.Socket):
+    async def _read(self, rcv_socket: zmq_async.Socket):
         t = time.time()
-        buffer = rcv_socket.recv(zmq.NOBLOCK)
+        buffer = await rcv_socket.recv(zmq.NOBLOCK)
         print(f'Rcvd {get_buffer_id(buffer)}, lat: {time.time() - t}')
         # TODO check if buffer_id exists to avoid duplicates, re-send ack on duplicate
         # TODO acquire buffer pool
         self._buffer_queue.append(buffer)
         channel_id = self._rcv_sockets[rcv_socket]
+        # TODO keep track of a low watermark, ack only if received buff_id is above
+        # TODO batch acks
+        # asyncio.run_coroutine_threadsafe(self._ack(channel_id, buffer), self._sender_event_loop)
+        asyncio.create_task(self._send_ack(channel_id, buffer))
 
-        # TODO keep track of a low watermark, schedule ack only if received buff_id is above
-        # schedule ack message for sender
+    async def _send_ack(self, channel_id: str, buffer: Buffer):
         buffer_id = get_buffer_id(buffer)
         ack_msg = AckMessage(buffer_id=buffer_id, channel_id=channel_id)
-        self._acks_queues[channel_id].append(ack_msg)
-
-    def _send_loop(self):
-        # TODO should we share logic with DataWriter's sender? This way we'll get in_flight limitation + cleaner thread?
-        while self.running:
-            for channel_id in self._acks_queues:
-                # TODO batch acks
-                ack_queue = self._acks_queues[channel_id]
-                batch_size = 10
-                if len(ack_queue) < batch_size:
-                    continue
-                acks = []
-                while len(ack_queue) != 0:
-                    ack_msg = ack_queue.pop()
-                    acks.append(ack_msg)
-                ack_msg_batch = AckMessageBatch(acks=acks)
-                self._send_acks(channel_id, ack_msg_batch)
-
-    def _send_acks(self, channel_id: str, ack_msg_batch: AckMessageBatch):
         send_socket = self._send_sockets[channel_id]
-        # TODO limit number of in-flights acks?
-        data = ack_msg_batch.ser()
-        bs = data.encode()
-        t = time.time()
 
-        # TODO handle exceptions, EAGAIN, etc., retries
-        # send_socket.send_string(data, zmq.NOBLOCK)
-        send_socket.send(bs)
-        for ack_msg in ack_msg_batch.acks:
-            print(f'sent ack {ack_msg.buffer_id}, lat: {time.time() - t}')
+        # TODO limit number of in-flights acks?
+        # TODO handle exceptions? retries?
+        data = ack_msg.ser()
+        # print(data)
+        # raise
+        t = time.time()
+        await send_socket.send_string(data, zmq.NOBLOCK)
+        print(f'sent ack {buffer_id}, lat: {time.time() - t}')
+
+    def start(self):
+        super().start()
+        # TODO wait for even loop to be set
+        time.sleep(0.01)
+        asyncio.run_coroutine_threadsafe(self._receive_loop(), self._receiver_event_loop)
+
+    def close(self):
+        super().close()
